Uploaded new refactored PROPRTS-1 data writer
Acts as a data writer program for PROPRTS 1 minute data collector. Writing data for each run to be appended external file to be used for later via another program. The data this program generates are a pair of coordinates where x would represent the time (unix Timestamp format) and y would represent the value of the cryptocurrency at that specific time.